requirements.txt
# -----------------------------
# Core App & Utilities
# -----------------------------
streamlit
requests
python-dotenv
torch
sentence-transformers

# -----------------------------
# Database & Vector Search
# -----------------------------
psycopg2-binary           # PostgreSQL connector
pgvector                  # Optional helper for pgvector types in Python

# -----------------------------
# LangChain Ecosystem
# -----------------------------
langchain
langchain-core
langchain-community
langchain-huggingface

# -----------------------------
# Optional Extras / Streamlit Components
# -----------------------------
streamlit-extras          # For switch_page_button and stylable_container

# -----------------------------
# Notes:
# - FAISS removed (no longer used)
# - We won't need langchain-postgres unless you migrate to LangChain PGVector
# - pgvector is optional but recommended if you handle vector types in Python
# -----------------------------
